% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_7DADM.R
\name{compute_7DADM}
\alias{compute_7DADM}
\title{Compute 7DADM}
\usage{
compute_7DADM(all_data)
}
\arguments{
\item{all_data}{}
}
\description{
For each period of seven consecutive days, this function computes the daily
max water temperature for each day, then computes the average of those daily
max temperatures. This implementation assumes the data has already been
quality-controlled to remove or impute data during time periods in which the
logger was dewatered, buried, or nonfunctional.
}
