% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_7DADM_and_exceedance.R
\name{compute_7DADM_and_exceedance}
\alias{compute_7DADM_and_exceedance}
\title{Compute 7DADM and identify site exceedances}
\usage{
compute_7DADM_and_exceedance(fall_data_loc, sum_data_loc, lookup_loc)
}
\arguments{
\item{fall_data_loc}{File path to directory with quality-controlled fall data.}

\item{sum_data_loc}{File path to directory with quality-controlled summer data.}

\item{lookup_loc}{File path to directory with two lookup tables called
site_group.csv and group_wqs.csv. Examples of these files are provided in the
lookup_tables folder; for other data beyond the example data, these need to
be edited to reflect the sites and water quality standards for the data of
interest.

Site-group lookup table:
\itemize{
\item Two sites are in the same group if they have the same WQS
schedule as each other throughout the water year
\item There is one row for each site
}

Example:\tabular{ll}{
   Site \tab Group \cr
   AndersonLower \tab 1 \cr
   AndersonUpper \tab 1 \cr
   CedarLower \tab 1 \cr
   CedarUpper \tab 2 \cr
}


Group-WQS lookup table:
\itemize{
\item Provides details for the WQS schedule for each group
\item There is one row for each group-WQS combination from 01/01 to 12/31
for example, if a group has three different WQS throughout
the water year, there are three rows for that group
\item This can be generated using the write_group_wqs() function,
or it can be written in a spreadsheet and exported to a csv file
called group_wqs.csv
}

Example:\tabular{llll}{
   Group \tab FromDate \tab ToDate \tab WQS \cr
   1 \tab 01/01 \tab 12/31 \tab 15 \cr
   2 \tab 01/01 \tab 02/14 \tab 12 \cr
   2 \tab 02/15 \tab 07/01 \tab 16 \cr
   2 \tab 07/02 \tab 12/31 \tab 12 \cr
}}
}
\description{
This function assumes the data come from a single water year with data in
two deployments, a fall deployment and a summer deployment. For example, for
the 2021-2022 water year as in the provided example data, the two deployment
seasons would be fall 2021 and summer 2022.
}
\examples{

fall_data_loc <- "data/2021_fall/4_final_data/"
sum_data_loc <- "data/2022_summer/4_final_data/"
lookup_loc <- "lookup_tables/"

compute_7DADM_and_exceedance(fall_data_loc, sum_data_loc, lookup_loc)

}
