% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataQCtools-package.R
\docType{package}
\name{dataQCtools-package}
\alias{dataQCtools}
\alias{dataQCtools-package}
\title{dataQCtools: Automate data cropping, QC, and more}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tribalxg.github.io/dataQCtools/}
}

}
\author{
\strong{Maintainer}: Jess Kunke \email{jessicakunke@gmail.com} (\href{https://orcid.org/0009-0006-1475-3734}{ORCID})

}
\keyword{internal}
