% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_qc_plots_of_cropped_data.R
\name{make_qc_plots_of_cropped_data}
\alias{make_qc_plots_of_cropped_data}
\title{Make QC plots of cropped data}
\usage{
make_qc_plots_of_cropped_data(cropped_loc, qc_plots_loc)
}
\arguments{
\item{cropped_loc}{File path to the directory containing the cropped data
generated by crop_raw_data() to use in making QC plots. These files must be
in csv format. The names of the cropped data files must start with
\if{html}{\out{<sitename>}}\emph{\if{html}{\out{<medium>}}}\if{html}{\out{<deployseason>}}_\if{html}{\out{<deployyear>}}. For example, a file with
air data from site NolanLower and deployment date Summer 2023 should have
a filename starting with NolanLower_air_sum_23, like
NolanLower_air_sum_23_cropped.csv.}

\item{qc_plots_loc}{File path to the directory where QC plots will be stored.}
}
\description{
This function is intended to be used on cropped data, such as the csv files created by crop_raw_data().
}
\examples{

cropped_loc <- "data/2022_summer/2_cropped_csv/"
qc_plots_loc <- "data/2022_summer/3_qc_plots/"

# get a list of all the cropped data files
filenames <- list.files(path = cropped_loc, pattern=".csv")
# this is now a list of all filenames; we haven't read in the data yet, but
# make sure this lists all the cropped data files you want to plot:
filenames

# then make QC plots
make_qc_plots_of_cropped_data(cropped_loc, qc_plots_loc)

}
