% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sites_to_WQS_groups.R
\name{match_sites_to_WQS_groups}
\alias{match_sites_to_WQS_groups}
\title{Match individual sites to water quality standards}
\usage{
match_sites_to_WQS_groups(wq_data, site_group)
}
\arguments{
\item{wq_data}{Data frame}

\item{site_group}{Site-group lookup table as an R data.frame object}
}
\value{
Data frame with site_group lookup table information
}
\description{
Use a site-group lookup table to determine which WQS group each site is in.
}
