% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_qcd_data.R
\name{read_in_qcd_data}
\alias{read_in_qcd_data}
\title{Read in quality-controlled data for all sites and deployments for a given water year}
\usage{
read_in_qcd_data(fall_data_loc, sum_data_loc)
}
\arguments{
\item{fall_data_loc}{File path to directory containing fall deployment data files; for water year 2021-2022, the fall deployment season is Fall 2021. Files must be csv format.}

\item{sum_data_loc}{File path to directory containing summer deployment data files; for water year 2021-2022, the summer deployment season is Summer 2022. Files must be csv format.}
}
\description{
Read in quality-controlled data for all sites and deployments for a given water year
}
