% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_impairments_table.R
\name{site_impairments_table}
\alias{site_impairments_table}
\title{Create a table of site impairments}
\usage{
site_impairments_table(wq_data)
}
\arguments{
\item{wq_data}{}
}
\value{
A table such as the following:\tabular{lll}{
   Site \tab Impaired \tab Dates of impairment \cr
   Site 1 \tab No \tab  \cr
   Site 2 \tab Yes \tab 3/17-3/19, 5/21 \cr
   Site 3 \tab Yes \tab 10/20, 4/28-4/29 \cr
   Site 4 \tab No \tab  \cr
   Site 5 \tab No \tab  \cr
   ... \tab ... \tab ... \cr
}
}
\description{
Given quality-controlled data for a given water year, this function
identifies which sites are impaired and the dates of their exceedances.
A creek/site is impaired if its temperatures exceed water quality standards
(WQS) at least twice during the water year and there is a period of at least
seven days between two of those exceedances.
}
