% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_group_wqs.R
\name{write_group_wqs}
\alias{write_group_wqs}
\title{Write group-WQS lookup table to file}
\usage{
write_group_wqs(outdir)
}
\arguments{
\item{outdir}{}
}
\description{
Write group-WQS lookup table to file
}
