% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_group_to_WQS.R
\name{match_group_to_WQS}
\alias{match_group_to_WQS}
\title{Match groups of sites to their corresponding water quality standards}
\usage{
match_group_to_WQS(wq_data, group_wqs)
}
\arguments{
\item{group_wqs}{}
}
\description{
Use group-WQS lookup table to determine what the WQS is for each group over
the course of the water year.
}
