% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_sites_to_WQS_groups.R
\name{match_sites_to_WQS_groups}
\alias{match_sites_to_WQS_groups}
\title{Match individual sites to water quality standards}
\usage{
match_sites_to_WQS_groups(wq_data, site_group)
}
\arguments{
\item{site_group}{}
}
\description{
Use a site-group lookup table to determine which WQS group each site is in.
}
