% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_qcd_data.R
\name{read_in_qcd_data}
\alias{read_in_qcd_data}
\title{Read in quality-controlled data for all sites and deployments for a given water year}
\usage{
read_in_qcd_data(fall_data_loc, sum_data_loc)
}
\arguments{
\item{sum_data_loc}{}
}
\description{
Read in quality-controlled data for all sites and deployments for a given water year
}
